refactor: move shell composition to workflow module
